+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## v0.1.0 - TBD
 
-+ Initial version of the `yayaml` module
++ Initial version of the `Yayaml` module